pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fasting-atlas"
version = "0.1.0"
description = "MVP parser for traceable extraction from human experiment papers"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7.0",
  "pdfplumber>=0.11.0",
  "requests>=2.32.0",
  "python-dotenv>=1.0.0",
  "pymupdf>=1.24.0",
  "pytesseract>=0.3.10",
  "pillow>=10.0.0",
  "opencv-python-headless>=4.9.0.80",
]

[project.scripts]
fasting-atlas = "fasting_atlas.cli:main"

[tool.setuptools.packages.find]
include = ["fasting_atlas*"]
